+++ Sources/SentryCrash/Recording/Tools/SentryCrashStackCursor_SelfThread.m
@@ -39,10 +39,35 @@ typedef struct {
     uintptr_t backtrace[0];
 } SelfThreadContext;
 
+static BOOL stitchSwiftAsync = NO;
+
+void
+sentrycrashsc_setSwiftAsyncStitching(bool enabled)
+{
+    stitchSwiftAsync = enabled;
+}
+
 void
 sentrycrashsc_initSelfThread(SentryCrashStackCursor *cursor, int skipEntries)
 {
     SelfThreadContext *context = (SelfThreadContext *)cursor->context;
+
+// backtrace_async api is only available from xcode 13
+#if __clang_major__ >= 13
+    int backtraceLength;
+    if (@available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, *)) {
+        if (stitchSwiftAsync) {
+            backtraceLength
+                = (int)backtrace_async((void **)context->backtrace, MAX_BACKTRACE_LENGTH, NULL);
+        } else {
+            backtraceLength = backtrace((void **)context->backtrace, MAX_BACKTRACE_LENGTH);
+        }
+    } else {
+        backtraceLength = backtrace((void **)context->backtrace, MAX_BACKTRACE_LENGTH);
+    }
+#else
     int backtraceLength = backtrace((void **)context->backtrace, MAX_BACKTRACE_LENGTH);
+#endif
+
     sentrycrashsc_initWithBacktrace(cursor, context->backtrace, backtraceLength, skipEntries + 1);
 }