+++ Tests/SentryTests/SentryCrash/SentryStacktraceBuilderTests.swift
@@ -67,7 +67,63 @@ class SentryStacktraceBuilderTests: XCTestCase {
         
         XCTAssertTrue(filteredFrames.count == 1, "The frames must be ordered from caller to callee, or oldest to youngest.")
     }
-     
+
+    func testConcurrentStacktraces() {
+        guard #available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *) else { return }
+
+        SentrySDK.start { options in
+            options.dsn = TestConstants.dsnAsString(username: "SentryStacktraceBuilderTests")
+            options.swiftAsyncStacktraces = true
+        }
+
+        let waitForAsyncToRun = expectation(description: "Wait async functions")
+        Task {
+            let filteredFrames = await self.firstFrame()
+            waitForAsyncToRun.fulfill()
+            XCTAssertGreaterThanOrEqual(filteredFrames, 3, "The Stacktrace must include the async callers.")
+        }
+        wait(for: [waitForAsyncToRun], timeout: 1)
+    }
+
+    func testConcurrentStacktraces_noStitching() {
+        guard #available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *) else { return }
+
+        SentrySDK.start { options in
+            options.dsn = TestConstants.dsnAsString(username: "SentryStacktraceBuilderTests")
+            options.swiftAsyncStacktraces = false
+        }
+
+        let waitForAsyncToRun = expectation(description: "Wait async functions")
+        Task {
+            let filteredFrames = await self.firstFrame()
+            waitForAsyncToRun.fulfill()
+            XCTAssertGreaterThanOrEqual(filteredFrames, 1, "The Stacktrace must have only one function.")
+        }
+        wait(for: [waitForAsyncToRun], timeout: 1)
+    }
+
+    @available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
+    func firstFrame() async -> Int {
+        return await innerFrame1()
+    }
+
+    @available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
+    func innerFrame1() async -> Int {
+        await Task { @MainActor in }.value
+        return await innerFrame2()
+    }
+
+    @available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
+    func innerFrame2() async -> Int {
+        let needed = ["firstFrame", "innerFrame1", "innerFrame2"]
+        let actual = fixture.sut.buildStacktraceForCurrentThreadAsyncUnsafe()!
+        let filteredFrames = actual.frames
+            .compactMap({ $0.function })
+            .filter { needed.contains(where: $0.contains) }
+        return filteredFrames.count
+
+    }
+
     func asyncFrame1(expect: XCTestExpectation) {
         fixture.queue.asyncAfter(deadline: DispatchTime.now()) {
             self.asyncFrame2(expect: expect)